use proper type for address and aligment
use ptr type to match 32 and 64 bit adress types.

source/odp_shared_memory.c: In function 'odp_shm_reserve':
source/odp_shared_memory.c:157:43: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
  addr = (void *)ODP_ALIGN_ROUNDUP_POWER_2((uint64_t)addr, align);
                                           ^
./odp_align.h:76:16: note: in definition of macro 'ODP_ALIGN_ROUNDUP_POWER_2'
  ((align) * (((x) + align - 1) / (align)))
                ^
source/odp_shared_memory.c:157:9: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
  addr = (void *)ODP_ALIGN_ROUNDUP_POWER_2((uint64_t)addr, align);

+++ arch/linux-generic/odp_shared_memory.h
@@ -58,7 +58,7 @@ extern "C" {
  *
  * @return Pointer to the reserved block, or NULL
  */
-void *odp_shm_reserve(const char *name, uint64_t size, uint64_t align);
+void *odp_shm_reserve(const char *name, uint64_t size, uintptr_t align);
 
 /**
  * Lookup for a block of shared memory
